PERA-1357 Refactor http client providers
