+++ wallet-sdk/src/androidMain/kotlin/com/algorand/common/foundation/network/HttpClientProvider.android.kt
@@ -10,33 +10,20 @@
  * limitations under the License
  */
 
-package com.algorand.common.foundation.network.algod
+package com.algorand.common.foundation.network
 
-import com.algorand.common.foundation.network.PeraJsonNegotiation
 import io.ktor.client.HttpClient
-import io.ktor.client.engine.darwin.Darwin
-import io.ktor.client.plugins.DefaultRequest
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.engine.okhttp.OkHttp
+import io.ktor.client.plugins.logging.ANDROID
 import io.ktor.client.plugins.logging.LogLevel
 import io.ktor.client.plugins.logging.Logger
 import io.ktor.client.plugins.logging.Logging
-import io.ktor.client.plugins.logging.SIMPLE
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
 
-internal actual fun getAlgodHttpClient(algodInterceptorPlugin: AlgodInterceptorPlugin): HttpClient {
-    return HttpClient(Darwin) {
+internal actual fun createHttpClient(): HttpClient {
+    return HttpClient(OkHttp) {
         install(Logging) {
-            logger = Logger.SIMPLE
+            logger = Logger.ANDROID
             level = LogLevel.BODY
         }
-        install(DefaultRequest) {
-            contentType(ContentType.Application.Json)
-        }
-        install(ContentNegotiation) {
-            json(PeraJsonNegotiation)
-        }
-        install(algodInterceptorPlugin)
     }
 }
