--- wallet-sdk/src/androidMain/kotlin/com/algorand/common/foundation/network/pera/PeraMobileHttpClientProvider.android.kt
@@ -1,42 +1,0 @@
-/*
- * Copyright 2022 Pera Wallet, LDA
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
- */
-
-package com.algorand.common.foundation.network.pera
-
-import com.algorand.common.foundation.network.PeraJsonNegotiation
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.okhttp.OkHttp
-import io.ktor.client.plugins.DefaultRequest
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.logging.ANDROID
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
-
-internal actual fun getPeraMobileHttpClient(peraMobileInterceptorPlugin: PeraMobileInterceptorPlugin): HttpClient {
-    return HttpClient(OkHttp) {
-        install(Logging) {
-            logger = Logger.ANDROID
-            level = LogLevel.BODY
-        }
-        install(DefaultRequest) {
-            contentType(ContentType.Application.Json)
-        }
-        install(ContentNegotiation) {
-            json(PeraJsonNegotiation)
-        }
-        install(peraMobileInterceptorPlugin)
-    }
-}
