+++ wallet-sdk/src/iosMain/kotlin/com/algorand/common/foundation/network/HttpClientProvider.ios.kt
@@ -10,33 +10,20 @@
  * limitations under the License
  */
 
-package com.algorand.common.foundation.network.pera
+package com.algorand.common.foundation.network
 
-import com.algorand.common.foundation.network.PeraJsonNegotiation
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.darwin.Darwin
-import io.ktor.client.plugins.DefaultRequest
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.logging.DEFAULT
 import io.ktor.client.plugins.logging.LogLevel
 import io.ktor.client.plugins.logging.Logger
 import io.ktor.client.plugins.logging.Logging
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
+import io.ktor.client.plugins.logging.SIMPLE
 
-internal actual fun getPeraMobileHttpClient(peraMobileInterceptorPlugin: PeraMobileInterceptorPlugin): HttpClient {
+internal actual fun createHttpClient(): HttpClient {
     return HttpClient(Darwin) {
         install(Logging) {
-            logger = Logger.DEFAULT
+            logger = Logger.SIMPLE
             level = LogLevel.BODY
         }
-        install(DefaultRequest) {
-            contentType(ContentType.Application.Json)
-        }
-        install(ContentNegotiation) {
-            json(PeraJsonNegotiation)
-        }
-        install(peraMobileInterceptorPlugin)
     }
 }
